Update devicePixelRatio in onMetricsChanged
Previously the devicePixelRatio was only set once at initialization,
when it possibly was not available yet. This could lead to a wrong
device type detection (e.g. a OnePlus Nord was reliably detected to be a
tablet when running a Flutter app in release mode). Since the
devicePixelRatio can change with an onMetricsChanged call, update the
static devicePixelRatio variable in that callback to fix the device type
detection.

Signed-off-by: Jim Gerth <jgerth@jusst.de>